+++ bigproject/src/main/java/org/bigseven/controller/FeedbackAdminController.java
@@ -4,8 +4,8 @@ package org.bigseven.controller;
 import jakarta.annotation.Resource;
 import jakarta.validation.Valid;
 import org.bigseven.constant.FeedbackStatusEnum;
-import org.bigseven.dto.AdminFeedbackRequest;
-import org.bigseven.dto.AdminFeedbackResponse;
+import org.bigseven.dto.Admin.AdminFeedbackRequest;
+import org.bigseven.dto.Admin.AdminFeedbackResponse;
 import org.bigseven.result.AjaxResult;
 import org.bigseven.service.FeedbackService;
 import org.springframework.web.bind.annotation.PostMapping;
