+++ bigproject/src/main/java/org/bigseven/dto/Admin/AdminFeedbackRequest.java
@@ -1,4 +1,4 @@
-package org.bigseven.dto;
+package org.bigseven.dto.Admin;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
