+++ bigproject/src/main/java/org/bigseven/dto/Admin/AdminFeedbackResponse.java
@@ -1,5 +1,4 @@
-package org.bigseven.dto;
-
+package org.bigseven.dto.Admin;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,11 +6,11 @@ import lombok.Setter;
 
 /**
  * @author v185v
- * &#064;date 2025/9/17
+ * &#064;date 2025/9/20
  */
 @Getter
 @Setter
 @AllArgsConstructor
-public class UserRegisterResponse {
+public class AdminFeedbackResponse {
     private Integer id;
 }
