+++ bigproject/src/main/java/org/bigseven/dto/Admin/AdminGetAllFeedbackRequest.java
@@ -1,0 +1,11 @@
+package org.bigseven.dto.Admin;
+
+import lombok.*;
+
+@Data
+@AllArgsConstructor
+@Builder
+public class AdminGetAllFeedbackRequest {
+
+    private Integer id;
+}
