+++ bigproject/src/main/java/org/bigseven/dto/Admin/AdminGetAllFeedbackResponse.java
@@ -1,0 +1,4 @@
+package org.bigseven.dto.Admin;
+
+public class AdminGetAllFeedbackResponse {
+}
