+++ bigproject/src/main/java/org/bigseven/dto/User/PublishFeedbackRequest.java
@@ -1,4 +1,4 @@
-package org.bigseven.dto;
+package org.bigseven.dto.User;
 
 
 import jakarta.validation.constraints.Size;
