+++ bigproject/src/main/java/org/bigseven/dto/User/UserRegisterRequest.java
@@ -1,4 +1,4 @@
-package org.bigseven.dto;
+package org.bigseven.dto.User;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
