+++ bigproject/src/main/java/org/bigseven/service/impl/FeedbackServiceImpl.java
@@ -1,5 +1,6 @@
 package org.bigseven.service.impl;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bigseven.config.FeedbackConfig;
@@ -14,6 +15,7 @@ import org.bigseven.exception.ApiException;
 import org.bigseven.mapper.FeedbackImageMapper;
 import org.bigseven.mapper.FeedbackMapper;
 import org.bigseven.mapper.UserMapper;
+import org.bigseven.result.AjaxResult;
 import org.bigseven.service.FeedbackService;
 import org.springframework.stereotype.Service;
 
@@ -32,6 +34,7 @@ public class FeedbackServiceImpl implements FeedbackService {
     private final UserMapper userMapper;
     private final FeedbackImageMapper feedbackImageMapper;
     private final FeedbackConfig feedbackConfig;
+    private final FeedbackService feedbackService;
 
     /**
      * 发布用户反馈信息
@@ -139,5 +142,23 @@ public class FeedbackServiceImpl implements FeedbackService {
             return;
         }
 
+    }
+
+    @Override
+    public void getAllFeedback(){
+        LambdaQueryWrapper<Feedback> feedbackQueryWrapper = new LambdaQueryWrapper<>();
+        feedbackQueryWrapper.orderByDesc(Feedback::getFeedbackId);
+        feedbackMapper.selectList(feedbackQueryWrapper);
+        List<Feedback> feedbackList = feedbackMapper.selectList(feedbackQueryWrapper);
+
+    }
+
+    @Override
+    public void getAllFeedbackByType(FeedbackTypeEnum feedbackType){
+        LambdaQueryWrapper<Feedback> feedbackQueryWrapper = new LambdaQueryWrapper<>();
+        feedbackQueryWrapper.eq(Feedback::getFeedbackType, feedbackType);
+        feedbackQueryWrapper.orderByDesc(Feedback::getFeedbackId);
+        feedbackMapper.selectList(feedbackQueryWrapper);
+        List<Feedback> feedbackListByType = feedbackMapper.selectList(feedbackQueryWrapper);
     }
 }
